♻️🚚Move purple submissions to submissions path
