+++ src/lib/server/types/purpleRegions.ts
@@ -1,0 +1,6 @@
+export enum PurpleRegion {
+	Aelio = 'aelio',
+	Retem = 'retem',
+	Kvaris = 'kvaris',
+	Stia = 'stia'
+}
